moved db to own package
