+++ server/firebasedb/database_test.go
@@ -1,0 +1,14 @@
+package firebasedb
+
+import (
+	"log"
+	"testing"
+)
+
+func TestConnectDB(t *testing.T) {
+	DBC.ConnectDB()
+
+	if DBC.DBCon == nil {
+		log.Fatal("DBClient not initialized correctly")
+	}
+}
